django/requirements.txt
Django
django-sitemaps
pillow
django-robots
markdown
whitenoise
gunicorn
